requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2.3
django-import-export==4.3.7
djangorestframework==3.16.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
openpyxl==3.1.5
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uritemplate==4.2.0
